05.04.17

+++ src/main/java/ru/settletale/math/Point.java
@@ -1,11 +1,11 @@
 package ru.settletale.math;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 public class Point {
-	Vector3f origin;
+	Vector3d origin;
 	
-	public Point(Vector3f origin) {
+	public Point(Vector3d origin) {
 		this.origin = origin;
 	}
 }
